Let kube-config be defined by KUBECONFIG env var
Helm does not pass the --kubeconfig flag to its plugins, but instead
sets the KUBECONFIG variable.  However the KUBECONFIG variable can be
a list of paths (colon-seperated).  At the same time, the methos from
the helm-plugin-utils package expect a single file when passed a
kubeconfig file parameter.  Therefore the KUBECONFIG variable cannot be
used as the value for field common.KubeConfig.File.

helm-plugin-utils parses the KUBECONFIG file properly as long as it
reads it from the environment.  Therefore, this commit simply leaves
common.KubeConfig.File empty, to have helm-plugin-utils process the
KUBECONFIG variable instead.

The user can still use Helm's --kubeconfig flag or use the KUBECONFIG
variable set in the user's environment.

Signed-off-by: Marc Khouzam <marc.khouzam@gmail.com>